Builtin commmands separated into different classes.
