+++ CLI/src/test/java/ru/spbau/sd/cli/interpreter/ParserTest.java
@@ -3,10 +3,7 @@ package ru.spbau.sd.cli.interpreter;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
-import ru.spbau.sd.cli.interpreter.commands.AssignmentCommand;
-import ru.spbau.sd.cli.interpreter.commands.BuiltinCommand;
-import ru.spbau.sd.cli.interpreter.commands.Command;
-import ru.spbau.sd.cli.interpreter.commands.ExternalCommand;
+import ru.spbau.sd.cli.interpreter.commands.*;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -14,12 +11,11 @@ import java.util.Collections;
 public class ParserTest {
     @Test
     public void builtinTest() {
-        final BuiltinCommand cmd = BuiltinCommand.cat;
         final String arg1 = "file.txt";
         final String arg2 = "software design.txt";
         Parser parser = new Parser(null);
-        ASTElement ret = parser.parseLine(String.format("%s %s '%s'", cmd, arg1, arg2));
-        Assert.assertEquals(BuiltinCommand.cat, ret.getCommand());
+        ASTElement ret = parser.parseLine(String.format("cat %s '%s'", arg1, arg2));
+        Assert.assertTrue(ret.getCommand() instanceof CmdCat);
         Assert.assertNull(ret.getPrevious());
         Assert.assertEquals(Arrays.asList(arg1, arg2), ret.getArguments());
     }
@@ -53,13 +49,13 @@ public class ParserTest {
         Parser parser = new Parser(null);
         ASTElement lastElement = parser.parseLine(pipeline);
 
-        Assert.assertSame(lastElement.getCommand(), BuiltinCommand.echo);
+        Assert.assertTrue(lastElement.getCommand() instanceof CmdEcho);
         Assert.assertEquals(lastElement.getArguments(), Collections.EMPTY_LIST);
 
         ASTElement firstElement = lastElement.getPrevious();
         Assert.assertNotNull(firstElement);
         Assert.assertNull(firstElement.getPrevious());
-        Assert.assertEquals(firstElement.getCommand(), BuiltinCommand.echo);
+        Assert.assertTrue(firstElement.getCommand() instanceof CmdEcho);
         Assert.assertEquals(firstElement.getArguments(),
                 Collections.singletonList(arg));
     }
